+++ doc/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.4.5 (dev)
+## 1.4.5
 
 - New sidebar menu, both the menu and the content are now a bit more responsive
 - Expose HEVA's plotly theme in app.py for ease of use